Improve time precision to milliseconds

+++ mysql-init/init-database.sql
@@ -1,6 +1,6 @@
 CREATE DATABASE gm1356_monitoring;
 USE gm1356_monitoring;
-CREATE TABLE readings (_value FLOAT(4,1), _timestamp INT);
+CREATE TABLE readings (_value FLOAT(4,1), _timestamp BIGINT);
 
 
 CREATE USER 'grafana'@'%' IDENTIFIED BY 'grafana';
